feat: make Logion reusable in other contracts.
logion-network/logion-internal#979
