+++ logics/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "logion_psp34"
+name = "logion_contract_pkg"
 version = "1.0.0"
 edition = "2021"
 authors = ["Logion Team <team@logion.network>"]
@@ -12,7 +12,6 @@ ink_prelude = { version = "4.3.0", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.6", default-features = false, features = ["derive"], optional = true }
 openbrush = { tag = "4.0.0-beta", git = "https://github.com/Brushfam/openbrush-contracts", default-features = false, features = ["psp34", "ownable"] }
-logion_contract_pkg = { path = "../../logics", default-features = false }
 
 [lib]
 path = "lib.rs"
@@ -26,6 +25,5 @@ std = [
     "scale/std",
     "scale-info/std",
     "openbrush/std",
-    "logion_contract_pkg/std",
 ]
 ink-as-dependency = []
