+++ logics/lib.rs
@@ -1,0 +1,4 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod impls;
+pub mod traits;
